- Fix unfinished sentence. (decff78057e18fd02f404fc3e4525c61b7466f33)
- Update lodepng_util.cpp to the latest version
(16e07418ad1bdb00211c2356bf62c7421098aefa)

Use double in zopfli_bin as we use doubles everywhere in Zopfli KrzYmod
which do not impact performance on x86 CPUs but ensure better
precission.

+++ src/zopfli/zopfli_bin.c
@@ -91,9 +91,9 @@ static void ZopfliInitBinOptions(ZopfliBinOptions* options) {
   options->legacy = 1;
 }
 
-static size_t ceilz(float num) {
+static size_t ceilz(double num) {
     size_t inum = (size_t)num;
-    if (num == (float)inum) {
+    if (num == (double)inum) {
         return inum;
     }
     return inum + 1;
@@ -443,7 +443,7 @@ static int Compress(ZopfliOptions* options, const ZopfliBinOptions* binoptions,
         if(binoptions->numblocks>fullsize) {
           i = 1;
         } else {
-          i = ceilz((float)fullsize / (float)binoptions->numblocks);
+          i = ceilz((double)fullsize / (double)binoptions->numblocks);
         }
         l=i;
         do {
@@ -656,7 +656,7 @@ static int Compress(ZopfliOptions* options, const ZopfliBinOptions* binoptions,
         if(binoptions->numblocks>fullsize) {
           i = 1;
         } else {
-          i = ceilz((float)fullsize / (float)binoptions->numblocks);
+          i = ceilz((double)fullsize / (double)binoptions->numblocks);
         }
         l=i;
         do {
